[formerly a2ea2c5fd6325b10b300de5d12bfbf9ebe3f232e] [formerly b6cac7d6ff571f49e8f72d3d8fe437a36ddc976e [formerly a77745bdb3e14aedb8a46ef3d5be7c575e48225c] [formerly b6cac7d6ff571f49e8f72d3d8fe437a36ddc976e [formerly a77745bdb3e14aedb8a46ef3d5be7c575e48225c] [formerly b6cac7d6ff571f49e8f72d3d8fe437a36ddc976e [formerly a77745bdb3e14aedb8a46ef3d5be7c575e48225c] [formerly ce78757f88c57bbf28eebc0ed8078881046fbecb [formerly 02ca6cabd40becbd124b0eb6fd3aa9b6c36ea4c1]]]] [formerly 19b3e4f8c3eab5f62f534f2d9b8184aefc8ce29b [formerly b1f6de8eabcba33e12156864ac0c911ad6095804] [formerly 19b3e4f8c3eab5f62f534f2d9b8184aefc8ce29b [formerly b1f6de8eabcba33e12156864ac0c911ad6095804] [formerly 19b3e4f8c3eab5f62f534f2d9b8184aefc8ce29b [formerly b1f6de8eabcba33e12156864ac0c911ad6095804] [formerly 340163ca598be5ab767333c08f4f564f7c2d72e9 [formerly 6ddd78dbb157c432d58ed5c6cb3e252aee9ea7c2]]]]]]




Former-commit-id: 550fd06e25fb9abb2ff626c05143e843692a6441 [formerly 9f5b198c656de4117ab723c4d3f6d492104a2470]
Former-commit-id: 7cb3d61baab69253af70828836bacb58b721bb4d [formerly 0745b802cb4b1a98d7e7a70766f2d257d7fd825b] [formerly 4f3e9c268771fcbc86bec0ce82d0a75d6f3fd1f1 [formerly 0722c99a563cddd235c0abca41198e550d094239]]
Former-commit-id: 298a74ccf7b4552740d32063af731f5698ae620e [formerly 9303dcc34b94a37ee0121b4f407dc98d2fbe4e22]
Former-commit-id: b9f508964e3873abbd51cc21e30e3d3657ab2f98

+++ plugins/inputs/trace/trace_zipkinv2.go
@@ -39,7 +39,7 @@ func (z *ZipkinTracer) parseZipkinJsonV2(octets []byte) error {
 		tAdpter.operationName = zs.Name
 
 		if zs.ParentID != nil {
-			tAdpter.parentID      = fmt.Sprintf("%x", *zs.ParentID)
+			tAdpter.parentID      = fmt.Sprintf("%x", uint64(*zs.ParentID))
 		}
 
 		if zs.TraceID.High != 0 {
@@ -48,7 +48,7 @@ func (z *ZipkinTracer) parseZipkinJsonV2(octets []byte) error {
 			tAdpter.traceID = fmt.Sprintf("%x", zs.TraceID.Low)
 		}
 
-		tAdpter.spanID        = fmt.Sprintf("%x", zs.ID)
+		tAdpter.spanID        = fmt.Sprintf("%x", uint64(zs.ID))
 
 		for tag, _ := range zs.Tags {
 			if tag == "error" {
